backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
pillow-heif==0.18.0
numpy==2.1.3
onnxruntime==1.20.1
watchfiles==0.24.0
aiohttp==3.10.5
aiomqtt==2.0.1
asyncua==1.1.0
